Primera versión del programa para persistir y publicar
